% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_gym.R
\name{get_pkmn_region}
\alias{get_pkmn_region}
\title{get_pkmn_region}
\usage{
get_pkmn_region(regiao = "")
}
\arguments{
\item{regiao}{The name of region}
}
\value{
The name of pokemon game used for collect information of gym leaders.
}
\description{
This is the function that collects the name of game of pokemon given the region of pokemon for collect information of gym leaders.
}
\references{
Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://www.serebii.net/.
}
