% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_type.R
\name{gym_type}
\alias{gym_type}
\title{gym_type}
\usage{
gym_type(tipo)
}
\arguments{
\item{tipo}{the pokemon type.}
}
\value{
The  dataframe of gym leader of certain type.

#' @references Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://www.serebii.net/.
}
\description{
This is the function that filter the all gym leaders of certain type that was collected by the function
poke_all_leaders.
}
