% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_gym.R
\name{poke_all_leaders}
\alias{poke_all_leaders}
\title{poke_all_leaders}
\usage{
poke_all_leaders()
}
\value{
The dataframe with basics information of all gym leaders  (name, type and city).
}
\description{
This is the function that collects the dataframes of
gym leaders (name, type and city) of all region.
}
\references{
Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://www.serebii.net/.
}
