% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_getdata.R
\name{poke_get_type}
\alias{poke_get_type}
\title{poke_get_type}
\usage{
poke_get_type(x)
}
\arguments{
\item{x}{The list of dataframes that contains informations of a pokemon.}
}
\value{
The array with the types of a pokemon.

#' @references Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/.
}
\description{
This is the function that collect the type of a pokemon.
}
