% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_getdata.R
\name{poke_info}
\alias{poke_info}
\title{poke_info}
\usage{
poke_info(pokemon, show_moves = FALSE)
}
\arguments{
\item{pokemon}{The name of pokemon}

\item{show_moves}{A boolean that allow the function make a list of moves of pokemon chosen.}
}
\value{
The list of dataframes with basics information of pokemons (types,base-status,height,weight,region) and
a list of moves(name,accuracy,damage class,power,description,pp and type) with show_moves = TRUE.

#' @references Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/.
}
\description{
This is the function that collects the basics information about a pokemon.
}
