% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_gym.R
\name{poke_leaders}
\alias{poke_leaders}
\title{poke_leaders}
\usage{
poke_leaders(geracao = "")
}
\arguments{
\item{geracao}{The name of a pokemon region.}
}
\value{
The dataframe with basics information of gym leaders  (name, type and city).
}
\description{
This is the function that collects the gym leaders (name, type and city) of a certain region.
}
\references{
Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://www.serebii.net/.
}
