% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_getdata.R
\name{poke_move}
\alias{poke_move}
\title{poke_move}
\usage{
poke_move(x, Movimentos)
}
\arguments{
\item{x}{The dataframe who contains information about a certain move.}

\item{Movimentos}{The name of move.}
}
\value{
The dataframe with basics information of pokemon move (name,accuracy,damage class,power,description,pp and type)
that are learned by level in Pokemon Ultra Sun & Ultra Moon.

#' @references Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/.
}
\description{
This is the function that collects the basics information about a certain Pokemon move.
}
