% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_getdata.R
\name{poke_move_info}
\alias{poke_move_info}
\title{poke_move_info}
\usage{
poke_move_info(move)
}
\arguments{
\item{move}{The name of a pokemon move.}
}
\value{
The dataframe with basics information of pokemon move (accuracy,damage class,power,description,pp and type).

#' @references Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/.
}
\description{
This is the function that collect the the especif  information about certain Pokemon move.
}
