% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_getdata.R
\name{poke_move_list}
\alias{poke_move_list}
\title{poke_move_list}
\usage{
poke_move_list(x)
}
\arguments{
\item{x}{The list of dataframes that contains informations of a pokemon.}
}
\value{
The dataframe with basics information of pokemon moves (name,accuracy,damage class,power,description,pp and type)
that are learned by level in Pokemon Ultra Sun & Ultra Moon.

#' @references Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/
}
\description{
This is the function that organize the basics information about all Pokemon moves filter by a function
poke_move.
}
