% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_trainer.R
\name{poke_print}
\alias{poke_print}
\title{poke_print}
\usage{
poke_print(dados)
}
\arguments{
\item{dados}{The list of dataframes of pokemon used by a trainer collected by poke_trainer.}
}
\value{
The html Table with the pokemon's used by a trainer.
}
\description{
This is the function that print the pokemon team.
}
\references{
Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/.
}
