% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_getdata.R
\name{poke_region}
\alias{poke_region}
\title{poke_region}
\usage{
poke_region(pokemon)
}
\arguments{
\item{pokemon}{The name of a pokemon.}
}
\value{
The  name of region of origin of a pokemon .

#' @references Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/.
}
\description{
This is the function that collect the region of
a certain pokemon.
}
