% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_getdata.R
\name{poke_status}
\alias{poke_status}
\title{poke_status}
\usage{
poke_status(x)
}
\arguments{
\item{x}{The list of dataframes that contains informations of a pokemon.}
}
\value{
The list of values of status and the name of status.

#' @references Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/.
}
\description{
This is the function that collect the base status (hp, attack, defense, special attack, special defense and speed) of
a certain pokemon.
}
