% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_trainer.R
\name{poke_trainer}
\alias{poke_trainer}
\title{poke_trainer}
\usage{
poke_trainer()
}
\value{
The list of dataframes with basics information of pokemons (types,base-status,height,weight,region and level).
}
\description{
This is the function that collects the  information about a pokemon team chosen from the user.
}
\references{
Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/.
}
