% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_type.R
\name{poke_type}
\alias{poke_type}
\title{poke_type}
\usage{
poke_type(tipo, show_moves = FALSE, show_trainers = FALSE)
}
\arguments{
\item{tipo}{the pokemon type.}

\item{show_moves}{A boolean that allow the function make a list of moves of pokemon chosen.}

\item{show_trainers}{A boolean that allow the function make a list of trainers of certain moves.}
}
\value{
The list of dataframes with basics information of  all pokemons (types,base-status,height,weight,region) of
certain type and a list of moves (name,accuracy,damage class,power,description,pp and type) with show_moves = TRUE.
}
\description{
This is the function that collects the  information about all pokemon of certain type.
}
\references{
Pokémon and Pokémon character names are trademarks of Nintendo. The data collect from https://pokeapi.co/
}
