% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water}
\alias{water}
\title{Dados sobre os pokemon do tipo Água}
\format{
Uma tabela com 13 colunas:
\describe{
\item{Nome}{Nome do Pokemon}
\item{Regiao}{Região do Pokemon localizado pela primeira vez.}
\item{Tipo Principal}{Tipo Principal do Pokemon}
\item{Tipo Secundario}{Tipo Secundario do Pokemon}
\item{hp}{O Hp base do Pokemon}
\item{attack}{O ataque base do Pokemon}
\item{defense}{A defesa base do Pokemon}
\item{special-attack}{O ataque especial base do Pokemon}
\item{special-defense}{A defesa especial base do Pokemon}
\item{speed}{A velocidade base do Pokemon}
\item{Peso}{O Peso base do Pokemon}
\item{Altura}{A Altura base do Pokemon}
\item{id}{O id dentro da pokemon dentro do pokeapi}
}
}
\source{
https://pokeapi.co/.
}
\usage{
water
}
\description{
A tabela, gerada a partir da função poke_type com o argumento 'water'
}
\keyword{datasets}
